Move some code to L0
