+++ sycl/plugins/unified_runtime/ur/ur.cpp
@@ -28,20 +28,3 @@ std::vector<ur_platform_handle_t> *PiPlatformsCache =
     new std::vector<ur_platform_handle_t>;
 SpinLock *PiPlatformsCacheMutex = new SpinLock;
 bool PiPlatformCachePopulated = false;
-
-// Global variables for ZER_EXT_RESULT_ADAPTER_SPECIFIC_ERROR
-thread_local ur_result_t ErrorMessageCode = UR_RESULT_SUCCESS;
-thread_local char ErrorMessage[MaxMessageSize];
-
-// Utility function for setting a message and warning
-[[maybe_unused]] void setErrorMessage(const char *message,
-                                      ur_result_t error_code) {
-  assert(strlen(message) <= MaxMessageSize);
-  strcpy(ErrorMessage, message);
-  ErrorMessageCode = error_code;
-}
-
-ur_result_t zerPluginGetLastError(char **message) {
-  *message = &ErrorMessage[0];
-  return ErrorMessageCode;
-}
